requirements.txt
appdirs==1.4.4
audioread==2.1.9
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
iopath==0.1.8
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
numba==0.53.1
numpy==1.20.2
olefile==0.46
opencv-python==4.5.2.54
packaging==20.9
Pillow==8.2.0
pooch==1.3.0
portalocker==2.3.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.22
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.6.2
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
tornado==6.1
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yacs==0.1.6
